Move AbstractEdgeLayout to AbstractEdge (from AbstractSimpleEdge)
